EDU-519 Move timeline component to interactive-media plugin
