+++ src/bootstrap/server-config.js
@@ -55,7 +55,6 @@ const configSchema = joi.object({
   additionalControllers: joi.array().items(joi.function().class()).default([]),
   consentCookieNamePrefix: joi.string().required(),
   areRoomsEnabled: joi.bool().default(false),
-  isAdminTestsTabEnabled: joi.bool().default(false),
   additionalHeadHtml: joi.string().default(''),
   plugins: joi.array().items(joi.string().required()).default(['markdown', 'image'])
 });
@@ -81,7 +80,6 @@ class ServerConfig {
       importSources: this.importSources.map(({ name, hostName, allowUnsecure }) => ({ name, hostName, allowUnsecure })),
       consentCookieNamePrefix: this.consentCookieNamePrefix,
       areRoomsEnabled: this.areRoomsEnabled,
-      isAdminTestsTabEnabled: this.isAdminTestsTabEnabled,
       plugins: this.plugins
     };
   }
