+++ src/components/pages/admin.js
@@ -1,16 +1,13 @@
 import { Tabs } from 'antd';
 import PropTypes from 'prop-types';
 import Restricted from '../restricted.js';
-import TestsTab from '../admin/tests-tab.js';
 import { useTranslation } from 'react-i18next';
 import cloneDeep from '../../utils/clone-deep.js';
 import SettingsTab from '../admin/settings-tab.js';
 import React, { useEffect, useState } from 'react';
 import { useBeforeunload } from 'react-beforeunload';
-import { useService } from '../container-context.js';
 import permissions from '../../domain/permissions.js';
 import StoragePlansTab from '../admin/storage-plans-tab.js';
-import ClientConfig from '../../bootstrap/client-config.js';
 import { confirmDiscardUnsavedChanges } from '../confirmation-dialogs.js';
 import { batchShape, documentMetadataShape, settingsShape, storagePlanWithAssignedUserCountShape } from '../../ui/default-prop-types.js';
 
@@ -18,7 +15,6 @@ const { TabPane } = Tabs;
 
 function Admin({ initialState, PageTemplate }) {
   const { t } = useTranslation('admin');
-  const clientConfig = useService(ClientConfig);
   const [currentTab, setCurrentTab] = useState('1');
   const [isCurrentTabDirty, setIsCurrentTabDirty] = useState(false);
   const [settings, setSettings] = useState(cloneDeep(initialState.settings));
@@ -71,11 +67,6 @@ function Admin({ initialState, PageTemplate }) {
                 onStoragePlansSaved={setStoragePlans}
                 />
             </TabPane>
-            {clientConfig.isAdminTestsTabEnabled && (
-              <TabPane className="Tabs-tabPane" tab={t('testsTabTitle')} key="100">
-                <TestsTab />
-              </TabPane>
-            )}
           </Tabs>
         </Restricted>
       </div>
