+++ src/components/pages/admin.yml
@@ -4,6 +4,3 @@ settingsTabTitle:
 storagePlansTabTitle:
   en: Storage plans
   de: Speicherpläne
-testsTabTitle:
-  en: Tests
-  de: Tests
